update/ split CardsSection & add GeneralButton appearance
